+++ tests/e2e/cases/fs_promises_stat_access_ignored_arguments/in.ts
@@ -1,5 +1,6 @@
 const events: string[] = [];
 const file = "/tmp/tsc2c-fs-promises-stat-access-ignored.txt";
+const defaultMode = undefined;
 
 function mark(label: string): number {
     events.push(label);
@@ -11,6 +12,7 @@ fs.writeFileSync(file, "stats");
 
 fs.promises.access(file, undefined, mark("access-default"));
 fs.promises.access(file, void 0, mark("access-void-default"));
+fs.promises.access(file, defaultMode, mark("access-alias-default"));
 fs.promises.access(file, fs.constants.F_OK, mark("access-mode"));
 
 fs.promises.stat(file, void 0, mark("stat")).then((value: FSStats): void => {